addition

+++ train.py
@@ -37,7 +37,7 @@ def greedy_decode(model, source, source_mask, tokenizer_tgt, max_len, device):
         if decoder_input.size(1) == max_len:
             break
         # build mask for target
-        decoder_mask = causal_mask(decoder_input.size(1)).type_as(source_mask).to(device)
+        decoder_mask = causal_mask(decoder_input.size(1)).long().to(device)
       
 
         # calculate output
@@ -49,7 +49,7 @@ def greedy_decode(model, source, source_mask, tokenizer_tgt, max_len, device):
         _, next_word = torch.max(prob, dim=1)
      
         decoder_input = torch.cat(
-            [decoder_input, torch.empty(1, 1).type_as(source).fill_(next_word.item()).to(device)], dim=1
+            [decoder_input, torch.empty(1, 1).long().fill_(next_word.item()).to(device)], dim=1
         )
 
         if next_word == eos_idx:
